skip the TERM/INT test

+++ scripts/open2.pl
@@ -59,11 +59,13 @@ my $EOL = $^O eq 'MSWin32' ? "\r\n" : "\n";
 
 #    is $r->open2({instr=>"first line${EOL}second line${EOL}"}, $^X, '-e', 'print join "", <STDIN>'), "first line${EOL}second line${EOL}", 'expected multi-line result';
 
+    if (0) { # seems to kill also the test script
 #    eval { $r->open2($^X, '-e', 'kill TERM => $$') };
     eval { $r->open2($^X, '-e', 'kill INT => $$') };
     like $@, qr{^Command died with signal 2, without coredump};
     is $@->{signalnum}, 2;
     is $@->{coredump}, 'without';
+    }
 
     eval { $r->open2($^X, '-e', 'kill KILL => $$') };
     like $@, qr{^Command died with signal $KILLrx, without coredump};
